Yesterday's hack to support parallel deps targets was bogus.  Can't have the libraries depending directly ontheir $(pretarg)'s.  That's a no no that I've seen before, and should've remembered.  It results in libraries getting rebuilt everytime a deps target is revalidated by make, which results in library builds everytime even if nothing changed.
So, this mod removes the dependence of the libraries upon $(pretarg).
In order to ensure that $(pretarg) is still built before the lib when
doing a parallel make, it is necessary to do something circuitous to
fool make into completing $(pretarg) before moving on to the library
target.  The only way I could find to do this was with an invocation
of a sub make.  This is done only in cases where $(pretarg) is
defined.  If it is, it is built as a dependent of the current target,
then the rule invokes a sub make which actually builds the library.