Initial commit of the Python C++ bidning machinery.  This is partly stuff from the Livermore effort, and partly stuff done here (which is in turn based on obsolescent portions of the Livermore package).  At this point we have resurrected the global function invocation trampoline.  Yet to be done is the member function invocation trampoline.
This revitalized implementation of the global function invocation
trampoline is dramatically improved over the prior implementation.  We
are now able to expunge all the hacky type registry nonsense.  This is
now fully replaced by glommalizing with PythonExtension<T>.  This
makes the process of building extensions fantastically simpler.

As before, the trampoline wraps the actual function invocation in a
try/catch block which prevents any C++ exception from percolating out
of the function and trashing the interpretter state.  Instead, any
exception which makes it to the trampoline handler is converted to a
Python exception.  There are three recognized cases:  A PyException,
which is an already registered Python exception, thrown by either the
invoked function, or some part of the C++ Python API that it calls, or
a char *, which is printed and registered as a Python run time error,
or any other C++ exception, which is registered as an unidentified
Python run time error.

In this way, you can safely call C++ functions from Python, without
fear of error conditions in the C++ code killing the interpretter.