Updated to new C++ standard syntax for template members which is enforced by KCC3.3, to use a nested template class we must use the following, new, syntax; 	typename MT::template CCSF<double> density;       // etc.... Since we do not have universal access to 3.3 yet we have changed OS_Mesh to include typedefs to the different Cell-centered nested classes, ie. 	typedef CCSF<double> CCSF_double; 	typedef CCVF<double> CCVF_double;       // etc.... we now use these in Opacity, Mat_State, and Tally.
A wonderfull side-effect of this KCC3.2 test, compiling with 3.2 produces
and executable which does not suffer the hideous 64-bit bus error that we
were experiencing under 3.2.  We now believe that this was definitely a
3.2 bug on the 64-bit machines dealing with vector<string> or some such
nonsense, we have messages into KAI to confirm this.  Anyway, all our
problems should be solved when we migrate fully to 3.3.