Added reference counting to Sprng.hh.  The Random Number libraries based on the SPRNG library are now officially, Rnd_Control and Sprng. Rnd_Control makes Sprng random number objects.  Sprng has reference counting so it can be copied and assigned without worry of freeing the random number state held by the SPRNG libraries.
The SPRNG libraries still have two (found) problems.  When compiling with
the lcg, there is a memory leak in spawn_rng.  When compiling with lfg,
there is a memory leak in init_rng (initialize).  These have been reported
to the library originators.  We will proceed with the lcg for now,
however, we will need the lfg because the lcg does not have enough streams
to accomodate our problems.