Add option shared=yes for building shared libs on SGI.  Note that you will probably need to set LD_LIBRARY64_PATH to make effective use of this.
I think it's probably a bad idea to try to mix shared and non shared
build products, so I've added a 'z' to the compiler mangling tag when
building shared components.  'z' is because many compilers have
historically used -z or +z to represent various forms of PIC
compilation modes.