Added a vertex-centered field (Mesh_XYZ::vctf<T>).
Added a class of small vectors (Mesh_XYZ::tiny_vec<T>).

Added Mesh_XYZ::get_fn() to get the normals on the cell faces.

Redesigned the scatter/gather operations in Mesh_XYZ.  The new and/or
improved operations are:
   gather from cctf to fcdtf;
   scatter from cctf to fcdtf;
   scatter from fcdtf to cctf;
   scatter from vctf to fcdtf.
The classes OpAddAssign and OpMultAssign were created as part of this
redesign.

Added several new tests in txyz.cc for the new functionalities.