Added to the P13T::calcMomentumDeposition() routine.  Now all of the sigma*F terms in the momentum deposition and velocity correction terms are calculated; we still need to add the sigma*phi terms.
Some basic functionality needs to be added/examined in the communications
and expression template libraries to make the job easier for P13T.
Specifically, we would like to be able to multiply scalar fields with
vector fields of the same centering without having to resort to iterators.
Also we need to be able to pass messages of non-basic types (such as
tiny_vecs): perhaps overload the update routines of guarded fields?