Changed the way that max_size() is calculated to be identical to KCC's implementation of the STL.  Note that our allocator class is not yet a true model of an Allocator as defined in Austern's book.  We will need to decide whether we should just use the one defined in <memory>.  In any case, this will be transparent to users of our Mat classes.
Additional note: because of compiler deficiencies we have two
workarounds in the code that tests whether the Mat1 class is a model
of a Random Access Container.  Because of the lack of two-phase name
binding in templates we must use the Mat classes with a type that has
a default constructor, even though that is not required by the Random
Access Container requirements.  Also, the compiler sometimes has
trouble converting a reverse_iterator to a const_reverse_iterator when
trying to resolve overloaded operators.  The current workaround in the
test code is to use a static_cast, but we may have trouble with
certain algorithms that don't do that (since they shouldn't have to).