Removed the census weight adjustment after combing and verified with the five full verification test problems.  A little history follows:
The census comb is essentially a splitting/Russian roulette technique,
and as such, it is more statistical than the old comb.  The new comb,
however, is reproducible, whereas the old one was not.  I then proceeded
to get cute and readjust the census energy-weights if there was no
unsampled census energy, thus exactly conserving energy.  I would not
do this readjustment if there was unsampled energy, which would occur
for too few particles or heterogeneous sources.  This selective adjustment
is actually biased--more so for smaller numbers of particles.  Moreover,
this adjustment can make the comb nonreproducible in parallel, especially
when using full domain decomposition.