This is the fix for bug report draco-build-system/28.  The build system can now build executables from DSO draco libraries.  The following caveats apply:
1) the rpaths only point to draco installed locations, moving these
libraries may cause the executable to fail;
2) any .so vendor libraries should be put in the users LD_LIBRARY_PATH so
that the executables can run; we do not (as of now) set rpaths to any
vendor libraries.

The basic fix consisted of these lines in ac_dracoenv.m4:

case sgi::

       # set rpath when building shared library executables
       if test "${enable_shared}" = yes; then
           DRACO_LIBS="-rpath \${libdir} ${DRACO_LIBS}"
           PKG_LIBS="-rpath .. ${PKG_LIBS}"
       fi

case sun::

       # set -R when building shared library executables
       if test "${enable_shared}" = yes; then
           DRACO_LIBS="-R \${libdir} ${DRACO_LIBS}"
           PKG_LIBS="-R .. ${PKG_LIBS}"
           RANLIB=':'
       fi

Notice that the sun requires deactivation of ranlib.  As mentioned in the
bug report, ranlib throws an error that causes make to fail.

We have piggy-backed the rpaths onto DRACO_LIBS for package executables
and PKG_LIBS for test executables.

These have been tested on the sgi and sun.