* Got rid of namelist.  If anyone is really attached to this, it should be easy to add it back in (at least easier than the removal). The constructor is now:
   PCG_Ctrl(const Method method = GMRS);

   where method is PCG_Ctrl::BAS, PCG_Ctrl::CG, or PCG_Ctrl::GMRS (see
   PCG Reference Manual for the meanings).
   Added various set* member functions to take the place of namelist.

* UEXACT (data member d_uExact) is allocated only if IPARM(IUEXAC) is
YES.  Since this is rarely used (you have to know the exact solution),
this should cut memory requirements by nearly a half.  See member
function setUexact.

* Most member functions renamed.  For example, `pcg_fe' is now `solve',
which now has the prototype

    void solve(dsxx::Mat1<T>& x,
               const dsxx::Mat1<T>& b,
               dsxx::SP< PCG_MatVec<T> > pcg_matvec,
               dsxx::SP< PCG_PreCond<T> > pcg_precond,
               const int nru = -1);

The only new parameter here is nru, which previously was specified in
the constructor.  If nru = -1, it's set from x.size(), otherwise
it must be <= x.size().

* Major re-write of implementation.