Added autodoc target to build system.  Now to build the autodocumentation using doxygen do the following (in the target directory):
1) configure draco.
2) do a gmake autodoc or enter the autodoc directory and run gmake.
3) a latex and html directory are built in the install location

gmake clean in the toplevel or autodoc directory will clean the html and
latex directory contents.  If configure cannot find doxygen, then any
attempt to run gmake autodoc will result in a warning message and exit.
Additionally, if autodoc is not checked out as part of draco, a warning
message is fired by gmake.