Added the Topology class.  Added testing for Topology in test.
We added a get_neighbors function to OS_Mesh in anticipation of a
Topology_Builder class.  The Topology_Builder class will exist at a
higher level than Topology because we do not know how packages will need
to decompose the mesh.

Topology class uses C4::node() and C4::nodes().  Thus we added c4 to the
dependency list for the mc package.