Added the Topology class.  Added testing for Topology in test.
We added a get_neighbors function to OS_Mesh in anticipation of a
Topology_Builder class.  The Topology_Builder class will exist at a
higher level than Topology because we do not know how packages will need
to decompose the mesh.

Topology class uses C4::node() and C4::nodes().  Thus we added c4 to the
dependency list for the mc package.

We made tst* parallel capable because mc is now parallel capable.

tstTopology only tests full replication Topologies (Of course this is the
simplest case:) ).  We need to add testing for full DD and DD/replication
topologies.

Finally, no functions exist yet in Topology to access boundary cell
information.  We will wait to test and add this interface capability
until we have a better idea on how it is to be used.  We anticipate this
being used by the communicator and communicator builder class at some point.