Broke the Topology class into an inheritance tree.  Now to add new, optimized topology classes simply inherit from Topology abstract base class.
Currently we have two derived classes: General_Topology - handles all
topology types but is bulky for replication; Rep_Topology - optimized for
replication topologies.

In the future we may wish to add a Local_Topology that only contains
information pertinent to a processor (instead of the whole spatial
topology).

The inheritance is public - so virtual polymorphism can be used.