Added specialization of the ContainerTraits for Mesh_XYZ::cctf and Mesh_XYZ::fcdtf.
Added optional argument of type T to swap_faces<T>(...).  This argument
specifies the value to use at the boundary when there is no value to be
swapped with.  It defaults to T(), which mimics its current use, but
now I can override that when I need a multiplicative identity instead of
the additive identity at the swapped boundary face.