Refactored the draco testing system.  Dejagnu is now bye-bye.  This makes the testing system much simpler.  The config files set the test_flags variable that determine the launchtest line.  In the makefile, the output from launchtest is piped through the test_filter and reports are generated.
As in the old test system, all of the executables and settings for a test
directory must be the same.  That is, all tests are parallel or all tests
are scalar, all tests are binary or all tests are python scripts, etc.  We
may change this in the future, however, for now we just provide the
functionality that we had before.

Note: we no longer need test/*.exp and test/*.0 files and directories in the
test subdirectories.  Dejagnu is no longer needed.