Reconfigured draco to take advantage of new build system communication stuff.  You can now use mpi on the LINUX machines (starting tomorrow when /usr/local/mpich gets ported).
To compile on linux with mpi set the environment variables as follows

	export MPI_INC_DIR=/usr/local/mpich/include
	export MPI_LIB_DIR=/usr/local/mpich/lib

and then configure

	configure (other options) --with-mpi-inc --with-mpi-lib
--with-c4=mpi

The SGIs work as before.  Note that --with-mpi is set to vendor by
default on SUNS and SGIs (same as before), but it is set to mpich by
default on linux. Thus, you don't have to explicitly say --with-mpi=mpich
on the linux boxes.  However, because the libraries and includes are not
in the standard path (as opposed to the SGIs) these must be specified
explicitly.