Updated configures in the HEAD REVISION of all packages.  This will not affect your tags because you should backtrack to config-1_2_0 to tag the releases.
This will have one affect.  When you commit configure using config-1_2_0 in
your packages, when you go back to the head revision your configure will
be out of date and you will have to autoreconf again.

For future notice, when we commit changes to the config directory, we should
do a global autoreconf and commit of draco.