Added new test to find the bug I just found in MeshXYZConnFacesAroundVertices.hh and MeshXYZConnFAVIterators.hh.
The bug was that the "CFAV" iterator assumed that the stride through
the vertex-centered field was one for the vertices of a cell and then,
of course, the stride of the cells would be eight.  Just the opposite
is the actual case.

I tested the tester with the old iterator, and it finds the bug.  The
new code passes the test.  Yippee!