Added some minor tests on local/global cell indices. ------------------------------------------------------------------------ Added functionality to exactly conserve energy after combing the census particles.
Specifically, in the base Source_Builder class, modified the comb to keep
track of numbers of particles; added a function, which is independent of
the topology, to reset the energy-weights of the particles in the census;
and added global-to-local and local-to-global cell conversions which are
critical for the full domain decomposition topology.

In the topology-dependent derived classes, added functions to recalculate
the census particles' energy-weights after combing and modified the
respective build_Source functions to use this new capability.

This new functionality will probably hose the regression tests....