Added Pack functionality to Topology inheritance tree.  The Pack nested classes are also an inheritance tree.  Thus we have:
	Topology::Pack (abstract base class)
	General_Topology::Pack
	Rep_Topology::Pack

The Pack interface has functions which (1) unpack and return a
SP<Topology> (2) return the pointer to the first position of the packed
data stream (3) return the size of the data stream (4) return an
indicator stating what type of topology this is.

These nested classes can be used for communication and writing to disk.