Added a constructor and accessor to Particle_Buffer:
	Particle_Buffer(int, Rnd_Control &);
	int get_inflight_dsize();

The new constructor allows the user to build a Particle_Buffer without
a constructed mesh.  Only the dimensionality is required.

The accessor returns the number of doubles for an inflight particle
(non-census).  The get_dsize() returns the size of census particles, which
is one less than inflight particles because census particles do not require
time_left fractions.

These have not been explicitly tested; but, they do not break any
existing tests.

QED