Tested the new non-default accumulators.
---------------------------------------------------------------------------
Modified several of the Tally class' accumulators to retain their
unity particle accumulation as a default but to allow accumulation
of multiple particles.  This change is transparent to current usage.
Also, added accessor functions.

Thus, we will now be able to take thread-private tally objects and,
using the global-tally accumulator functions with the thread-private
tally accessor functions, collapse them to a thread-global tally.