Changed mod_with_2e9 to INTEGER_MODULO_1E9.
This is now consistent with the default Rnd_Control argument of 1e9
streams.  Changing to 2e9 affects the spawn random number streams (this
may be a sprng bug) .  This is ok but it would change our downstream
regression tests which we don't want to do for this release.

We have "advanced" these commits to imc-2_2_0.

QED