Added AC_BUILD_CONFIGURES macro.  This can be used in any directory ( other than the toplevel draco directory).  It is used to build configure scripts; thus we only store one configure script in the repository:
	draco/configure

The configure.in in draco contains the following:

	AC_BUILD_CONFIGURES(.) --> the argument is the relative location
				   of the draco/config directory

In draco/src/configure.in we have the following:

	AC_BUILD_CONFIGURES(..)

In draco/doc/configure.in we have:

	AC_BUILD_CONFIGURES(..)

In draco/doc/MatrixFactory/configure.in we have:

	AC_BUILD_CONFIGURES(../..)

etc.

This architecture should break the package release dependence upon the
build system (specifically with respect to the configure scripts as
there need be only one configure script in the repository).

Of course, in the course of development you are free to use
autoreconf --localdir=config under draco/ to build configures locally.  I am
adding .cvsignore files to the directories that will cause cvs to ignore
configure files.

QED