Modified the CDI tests so that they no longer use UnitTestFrame.  This is not complete yet.  Still have to fix tCDI and finish tDummyOpacity (the multigroup part).
This is a good example of using the new elisp functions
draco-package-test and draco-cc-test.  Each test (tDummyEoS,
tDummyOpacity, etc) is a stand-alone executable.  The dummy files
(DummyEoS, DummyGrayOpacity, etc) are compiled in the libcdi_test once.
They are linked into the executables.  See Makefile.target for details.
Most of this is setup automatically.  The only thing that is required is
to add the dependencies tDummyEoS.o, etc to Makefile.target.

This testing environment provides tons of freedom.  The user can
basically do whatever he/she wants provided the tests produce output of
the form "test: pass/fail". Of course, the default setup works quite
nicely.

QED