Added Plank integration routines to CDI. These routines are static functions that calculate the normalized Planck integral, b(x) where
	b(x) = 15/pi^4 * x^3/(e^x - 1)

The routines do the following three operations:

1) given a frequency and temperature (keV) they return the integral, this
is always available (no multigroup boundaries have to be defined).

2) given a group index and temperature it returns the integral over the
group;  the multigroup boundaries must be set to use this (calling
setMultigroupOpacity for any CDI object)

3) integrate over all groups (not yet implemented); as above, the
multigroup boundaries must be set

CDI has been modified to check for that the group boundaries of all
multigroup opacities set in CDI are the same.  The group boundaries are
stored as static data in the class so they may be accessed globally
through

	CDI::getFrequencyGroupBoundaries()

Assertions are thrown if a multigroup opacity is set in ANY cdi object
with different frequency groups.

QED