Added virtual function:
		virtual std::vector<char> pack() const = 0;

	to GrayOpacity, MultigroupOpacity, and EoS.  Thus, derived types will
be able to pack themselves into char strings for persistence.  The drawback
is that the client must know what derived type of GrayOpacity, etc class is
packed to unpack them because the vector<char> should be given to a
constructor of the appropriate derived type.

	For example, assume you have a GrayOpacity that is a base class
pointer to Analytic_Gray_Opacity, you pack up the data as follows...

	//...gopac is a GrayOpacity pointer to an Analtyic_Gray_Opacity object

	vector<char> packed = gopac->pack();

	// now do stuff
	// ....

	// now later we want to unpack it
	GrayOpacity *ago = new Analtyic_Gray_Opacity(packed);
	// see, we had to know that the packed data object actually
	// referred to an Analytic_Gray_Opacity; in other words, there
	// is no way to call GrayOpacity ago(packed);

QED