Modified the behavior of the Random Walk procedure such that the sphere is always slightly less than the orthogonal distance to boundary, thus avoiding particle tracking issues with a coincident sphere and cell boundary.  The reduction is equal to 1.0e-6 of the smallest dimension of the cell.
For the OS_Mesh, the get_orthogonal_dist_to_bnd function is wrapped by
the get_random_walk_sphere_radius function, the latter of which does the
reduction.

For the RZWedge_Mesh, all these functions are specific for not
constraining the random walk sphere by the reflecting y-faces of the
RZWedge.  Therefore, the get_orthogonal_dist_to_bnd function has been
replaced with the get_random_walk_sphere_radius function which performs
the dist-to-bnd in the x- and z-dimensions and the reduction.

Renamed sample_pos_on_sphere to sample_random_walk_sphere in each of
OS_Mesh and RZWedge_Mesh because the returned position and outward normal
had different interpretations for each mesh type, and the function name
was simply wrong for the RZWedge_Mesh.  Mnemonics are us.

Updated the function names accordingly.