Added --enable-dlopen option to build system.  This is on by default if --enable-shared is on, and off by default, otherwise.
shared_lib: Added static function Shared_Lib::is_supported(), which
returns true if the platform is supported.  However, even if the
platform is unsupported, the all Shared_Lib code will compile and
link.  This allows portable code without the requirement for using
macros by the client.

config/Makefile.test.in: Instead of looking only in srcdir for
Makefile.target, first in the build directory, then look in the
srcdir.  This way, one can define a Makefile.target.in, so that a
Makefile.target gets built in the build directory.