We have changed the template compilation option on Q from -timplicit_local to -pt (the default).  This option specifies the followign behavior:
  -pt  Automatically instantiate templates into the repository with
       external linkage. Manually instantiated templates are placed
       in the output object with external linkage. This option is the
       default.

A side-effect of the -timplicit_local option is that multiple external
symbols can be defined that have different state.  This was seen in
uncleMcFlux.  The -pt option resolves this discrepancy.

We have one issue in to the compaq engineers about this option.  In
wedgehog a runtime symbol error was observed.  The missing symbol was a
virtual table.  We resolved the error by defining the destructor in the
derived class.  The error message was as follows:

The virtual table
'__vtbl_7rtt_imc74Flat_Mat_State_Builder__TQ16rtt_mc12RZWedge_MeshQ17rtt_imc14Gray_Frequency__7rtt_imc'
for class 'rtt_imc::Flat_Mat_State_Builder<rtt_mc::RZWedge_Mesh,
rtt_imc::Gray_Frequency>' is undefined, make sure that the first
noninlined virtual member function is defined.

I believe this to be an error because the symbol is getting built in the
cxx_repository; it just isn't getting put into the library.  (However,
when it is templated on other types there are no problems; hence, a bug).
In any case, if you have symbol problems they can probably be rectified
by defining destructors in the derived classes (because the destructor is
usually the first virtual member).

Tracker: 6048
Status: Open[ ] Closed[x] Pending[ ]