Changes necessary to run the regression tests on Lambda
The regression tests were failing on Lambda for a convoluted
reason. Lambda requires all mpirun calls to be prefaced with a call to
a script 'mpijob' which apparently creates a machines file for mpirun
to use. This script also kills runaway jobs. When executed from the
command line, it works fine, when executed from within the makefile
for regression testing is was prematurely killing jobs which it deemed
to be runaways.

A modifed (read: hacked) version of mpijob has been added to libexec
and launchtest has been modified to use it.


Tracker:
Status: Open[ ] Closed[ ] Pending[ ]