Chanaged the name of my checout tool from checkout.py to just   checkout. It can be invoked without python so this name is more   natural.
New feature is the ability to specify components within a package to
  checkout, e.g. draco/tools, and to spcify a destination directory,
  e.g.:
    > checkout -d tools draco/tools

Also added jayenne tools to the list of packages which can be checked out.

See the help file for more info and examples.

I'll move this script to environment/bin when I figure out how to
  portably locate python modules across platforms and environments.



Tracker:
Status: Open[ ] Closed[ ] Pending[ ]