- Fixed bug in Thin_Ptr.  The following now compiles under g++ when DBC is off:             DBC_Ptr<double> foo(new double);             DBC_Ptr<double> bar;             bar = foo; Fixing this required adding another ctor.
- Fixed memory leak in Safe_Ptr.  The reference counter was typically
being allocated twice in the pointer ctor.

- Added tests for when DBC is off (nothing can be caught, but it
should comile and run).  More tests are probably needed in test_basic.

- Delete memory in test so that memory checkers (e.g. valgrind) don't
print so many "problems."  Without this, it was hard to find the
memory leak in Safe_Ptr.  There are still two memory leaks in the test
(from the reference counter, DBC on only), but these are OK, because
they only occur when exceptions are thrown.