Removed libg2c checks from configure/build system.  This was done for the following reasons:
1) new versions of gcc (gcc > 3.4.5) do not include g77/g2c anymore, thus the build
system was failing whenever we tried to port to a gcc version greater than 3.4.5.
Since gcc 4.1.1 is now out we really need to get past this.

2) g2c is not necessarily needed for gandolf, lapack, etc.  It is only needed if those
libraries are built with g77.  There is no easy way to check how the vendor was built.


So, if you need to link in g2c you need to add it manually.  The build system already
provides a mechanism for doing this: --with-libs.  On our linux platforms one would
add the following:

	./configure --with-libs='-L/usr/lib/gcc/i386-redhat-linux/3.4.5 -lg2c'

This will append the g2c linkage to the LIBS variable (after all other vendors have
been added).