Added a function to return the vector of pairs that maps between the sequential moment index used in the moment-to-discrete operator and the (l,m) index of the spherical harmonics.
This is used in the construction a vector of scattering coefficients,
sigma_0, ..., sigma_{L-1} for use in the new anisotropic scattering operator
Coherent_Scattering. This new operator is to be distinguished from the Coherent_Isotropic
scattering operator, which simply multiplies the (0,0) moment by sigma_0 and any other
additional flux moments by zero.