+ Adding a new "unique" number accessor for LFG and LFG accessories.
The "get_unique_num()" accessor provides a combination of the RNG stream
ID and its state.  We (Jayenne Project) were previously relying on
get_num() to provide some degree of uniqueness, but this doesn't work
on particles that were previously split, as they share the same
generator stream ID. See
https://tf.lanl.gov/sf/go/artf23409?nav=1&_pagenum=1&returnUrlKey=1333470445809
for further details. This "unique" number can probably be made to
break if pushed to extreme limits, but it should be "unique enough"
for testing (One of the tests checks for uniqueness after each
of 10^6 calls to the RNG).

I've also added some simple tests to extend code coverage and removed tabs.