+ Adding a test showing how to use derived types with iso_c_binding
This is a example that shows how to provide a Fortran module that properly
defines a C interface to a host code including the use of (simple) derived
types (structs on the C side).  The idea here is that the host code could "use"
a module that we create and release in order to prevent the usual memory errors
that tend to arise when the Fortran host doesn't know anything about the
interface to the C function, like variables of the wrong type and arguments in the
wrong order.

Additionally, centralizing related data in derived types could reduce the
development impact of changing flat interfaces in multiple regression tests, as
the regression tests can also "use" the same module and automatically bring in
any changes to the derived type.
 
This coding requires the F2003 standard "import" and "iso_c_binding" intrinsics
which currently appear to be broken for character streams on PGI 10.4 or less.