+ Testing an F2C enumeration ability.
Unfortunately, F2003 enumerations are useless.  Enumerations are provided, but
enumeration *types* are not. So they are basically glorified parameter
statements that interpret special enumeration variables as c_ints.  However,
enumerations are still useful to us on the "C" side of the interface, so it's
worth trying this out on multiple platforms/compilers.    
http://www.rhinocerus.net/forum/lang-fortran/92750-no-enumeration-data-type-fortran.html