Add Random123 known-answer and unit tests to rng/test
Random123 comes with a collection of examples, several of which are
actually known-answer tests or unit tests.  The Random123
documentation strongly recommends that these tests be run on the
target system, with the same compiler flags as any code that will use
Random123, and the most reliable, straightforward way to do that is to
incorporate them into the rng package as rng tests.

This commit simply includes several of the Random123 examples,
unchanged, which means that they don't report success or failure the
way CTest expects.  A subsequent commit will update them to emit
PASSED and FAILED messages.

These will be built only if Random123 was found.