Set -ftz for all build types.
We don't seem to depend on denormals.  I removed -ftz in commit 6821
on 10/12/2012, believing that it was enabled by default for any value
of -O other than -O0, but some Intel compiler documentation (in
particular, "Consistency of Floating-Point Results using the Intel
Compiler") points out that -fp-model precise also disables -ftz, in
the interest of value safety.