extend c4 to begin supporting MPI user-defined vector types.
create_data_vector is basically a wrapper for the MPI_Type_vector and MPI_Type_commit functions. It creates an MPI user defined vector type.

I've typedef'd C4_Datatype to be MPI_Datatype unless we're in scalar build, in which case C4_Database is typedef'd as void * (these functions are all no-ops in scalar mode.)

send (with an added C4_Database argument) and receive (with an added C4_Database argument) use the type for actual blocked send and receive.

test/tstVector_Type.cc illustrates the use of these functions.