When using OpenMPI, set --mca mpi_paffinity_alone 0 for MPI-only tests.
The --bind-to-none option available (and the default) in OpenMPI 1.4+
is apparently not a synonym for this;

% ompi_info -param mpi all

with OpenMPI 1.6.3 on Moonlight reports that mpi_paffinity_alone is 1,
which means, "[A]ssume that this job is the only (set of) process(es)
running on each node and bind processes to processors, starting with
processor ID 0."  Setting mpi_paffinity_alone to 0 allows parallel
ctest to work correctly.  MPIEXEC_POSTFLAGS only affects MPI-only
tests (and not MPI+OpenMP tests).