+ This is the initial commit provided by Kent.  Here are his notes:
I have rewritten SP to conform to the specification for shared_ptr in
C++11. I have not implemented everything in that standard (some cannot
be implemented without language features introduced in C++11) but what
has been implemented has been tweaked to conform, and what goes beyond
the standard has been removed.

It all works now in Draco and passes benchmarks, albeit with some
benchmarks in ds++/tstSP tweaked to reflect the requirements of the
standard. I will hold off checking this in until I have gotten some
further experience bringing Capsaicin into line and have given y'all a
chance to absorb the implications.

Here is what I've changed in SP to make it closely resemble
shared_ptr:

1. The shared_ptr standard distinguishes empty pointers and null
pointers and imposes some requirements for empty pointers in some
places. This is not really visible to the average user of SP, being
mostly a matter of not allocating an internal reference count if the
pointer is NULL, but I mention it for completeness. This probably very
slightly decreases memory usage and is probably a wash for speed.

2. There is no silent dynamic casting any more. This is probably the
most visible and, in almost every way, most welcome change. A lot of
goofs will be found at compile time instead of at run time now.

3. I've implemented the dynamic_pointer_cast template for those cases
where you really have to have a dynamic cast. I've not found many so
far. I suspect more will crop up as I get into Capsaicin.

4. The standard shared_ptr does not provide a direct assignment from
raw pointer to shared_ptr. This is probably the change that will have
you gritting your teeth and banging your heads on your desks the
most. But there are good reasons it is the standard and we might as
well get out of the habit of making such assignments now.

The alternatives provided by the C++11 standard are: Do an explicit
cast to SP as part of the assignment, e.g.,

    my_smart_ptr = SP<my_class>(my_raw_pointer);

Or use the reset function:
    my_smart_ptr.reset(my_raw_pointer);

Or use the make_shared template:
    my_smart_pointer = make_shared<my_class>(my_constructor_arg);

The last only works for construction, but this is a very common case,
and for this case it's nifty, elegant, safe, and impossible to
implement generally without C++11 language features. (Sigh.) I've
mostly gone with using reset to replace the now-forbidden direct
assignment.  5. The bp() function is now spelled get(). It already had
the same functionality as the standard, just spelled differently.

6. There is no longer an operator!. This is redundant with operator
bool and is not in the standard.

7. Only the operator== and operator< relations are defined for
comparing and ordering smart pointers. I've yet to see this make a
difference; there are templates in namespace std, widely introduced
with one header or another, that specifies how to construct all
possible relationships out of these two.

The standard calls for relationships operators making comparisons to
nullptr. I have not bothered to implement this type and have therefore
not implemented these comparisons. The chief place this bites us is in
the common idiom, my_smart_pointer != NULL, which can simply be
replaced with my_smart_pointer (converted implicitly to bool.)

8. Kind of implied by my comments under (4), but reset is implemented
for both the no argument and the single raw pointer argument
signature. I haven't bothered with the other signatures; not vitally
needed.

refs #457