Attempt to stand up a style check system within travis-ci.
The goal of this change is to add an extra check for PRs that compares modified
code against the .clang-format ruleset.  If nonconformances are found, these are
printed as a diff and the style check result is 'fail.'  Users can run
'regression/check_style.sh' manually from their work directory to update code to
match the style requirements (git-clang-format must be installed locally for
this to work).

+ Add a new shell script, check_style.sh, that will run git-clang-format.
+ Remove some of the rules from .clang-format.
+ Update .travis.yml so that clang-format-3.8 is installed (hopefully) and add
  STYLE as a matrix option. Also, run the check_style.sh script after the
  coverate test is complete.
+ fixes #4