Add an Abstract_Class_Contexual_Parser that parallels Abstract_Class_Parser but with parser context.
The class Abstract_Class_Contextual_Parser is nearly identical to Abstract_Class_Parser; the only significant difference is that the Parse_Function typedef is to a function taking an additional context argument. Its use is documented in the test file test/tstAbstract_Class_Contextual_Parser.cc

I noticed while writing this class that there was an uncovered destructor in Abstract_Class_Parser_Base. On further reflection, I realized that this class is much more machinery than is needed to perform its function, so I stripped it down to its core and moved this into Abstract_Class_Parser.i.hh.